Move fake extensions clienset to own package
