+++ handler/routing_test.go
@@ -15,7 +15,7 @@ import (
 
 	"github.com/dgrijalva/jwt-go"
 	"github.com/gorilla/mux"
-	"github.com/kubermatic/api/extensions"
+	"github.com/kubermatic/api/extensions/fake"
 	"github.com/kubermatic/api/provider"
 	"github.com/kubermatic/api/provider/cloud"
 	"github.com/kubermatic/api/provider/kubernetes"
@@ -43,7 +43,7 @@ func createTestEndpoint() http.Handler {
 	kps["master"] = kubernetes.NewKubernetesFakeProvider("master", cps)
 
 	router := mux.NewRouter()
-	routing := NewRouting(ctx, dcs, kps, cps, true, base64.URLEncoding.EncodeToString([]byte(jwtSecret)), extensions.FakeWrapClientsetWithExtensions(nil))
+	routing := NewRouting(ctx, dcs, kps, cps, true, base64.URLEncoding.EncodeToString([]byte(jwtSecret)), fake.FakeWrapClientsetWithExtensions())
 	routing.Register(router)
 
 	return router
